+++ gulp/buildTask.js
@@ -5,16 +5,6 @@
  */
 module.exports = (gulp, $, config) => {
 
-    gulp.task('compile' , () => {
-        $.runSequence( 'clean' , 'sp' , 'html' , 'images' , 'sass' , 'js' )
-    });
-
-    gulp.task('local' , () => {
-		return gulp
-        .src('src/**/*.html')
-		.pipe($.replace('../../../dist', '/dist'))
-		.pipe($.replace('../../../src', '/src'))
-		.pipe(gulp.dest('src'))
-    });
+    gulp.task('build', gulp.series('clean' , gulp.parallel('js', 'html' ,'images'), 'spsass' , 'purify' ));
 
 };