+++ gulp/docTask.js
@@ -1,3 +1,7 @@
+const sassdoc       = require('sassdoc'),
+      del           = require('del'),
+      vinyl         = require('vinyl-buffer');
+
 /**
  * @param gulp
  * @param $
@@ -5,22 +9,26 @@
  */
 module.exports = (gulp, $, config) => {
 
-    const sassdoc = require('sassdoc');
+    function cleandoc(){
+        return del('docs')
+    }
+    cleandoc.description = 'scss guide 문서 폴더를 지웁니다.'
 
-    gulp.task('doc', ['cleandoc'] , () => {
+    function scssdoc() {
         const options = {
             dest: './docs'
         }
         return gulp
         .src(config.scss.src)
-        .pipe($.vinylBuffer())
+        .pipe(vinyl())
         .pipe(sassdoc(options))
         .resume();
-    });
+    }
+    scssdoc.description = 'scss guide 문서를 생성합니다.'
 
+    gulp.task(cleandoc);
+    gulp.task(scssdoc);
 
-    gulp.task('cleandoc', () => {
-        return $.del.sync('docs');
-    });
+    gulp.task('doc', gulp.series('cleandoc' , 'scssdoc'))
 
 };